Changed structure of project
